codigos/protocolos-api-python/grpc_api/requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf>=4.21.6,<5.0dev
